+++ cli/src/main/java/org/ehrbase/cli/config/CliOverwriteConfig.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2019-2024 vitasystems GmbH.
+ *
+ * This file is part of project EHRbase
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ehrbase.cli.config;
+
+import org.ehrbase.configuration.config.validation.ValidationConfiguration;
+import org.ehrbase.openehr.sdk.validation.terminology.ExternalTerminologyValidation;
+import org.springframework.boot.autoconfigure.flyway.FlywayMigrationStrategy;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+@Configuration
+public class CliOverwriteConfig {
+
+    @Bean
+    @Primary
+    public FlywayMigrationStrategy flywayMigrationStrategy() {
+        return flyway -> {
+            // Nop - prevent any flyway interaction
+        };
+    }
+
+    @Bean
+    public ExternalTerminologyValidation externalTerminologyValidator() {
+        return ValidationConfiguration.nopTerminologyValidation();
+    }
+}