+++ CHANGELOG.md
@@ -7,7 +7,36 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
  ### Added
+* Feature toggle for ehrscape API (ehrbase.rest.ehrscape.enabled, default: false) ([#1415](https://github.com/ehrbase/ehrbase/pull/1415))
+* Add experimental `AQL` support for `FOLDER` ([#1401](https://github.com/ehrbase/ehrbase/pull/1401))
+ ### Changed
+ ### Fixed 
+
+## [2.9.0]
+ ### Added
+ ### Changed 
+ ### Fixed
+* Improved transaction-awareness of caches ([#1407](https://github.com/ehrbase/ehrbase/pull/1407))
+
+## [2.8.1]
+ ### Added
+ ### Changed 
+ ### Fixed 
+* Fixed HTTP 500 in case stored query does not exist [#1409](https://github.com/ehrbase/ehrbase/pull/1409)
+
+## [2.8.0]
+ ### Added
+* AQL: Support query parameter with multiple values ([1399](https://github.com/ehrbase/ehrbase/pull/1399))
+ ### Changed 
+* Enable user access to the welcome page while authentication is enabled ([#1400](https://github.com/ehrbase/ehrbase/pull/1400))
+ ### Fixed 
+* Stored AQL queries: Correctly clean up query cache when storing queries ([#1405](https://github.com/ehrbase/ehrbase/pull/1405))
+* Updating an `EHR_STATUS` or `FOLDER` did not check the `If-Match header` ([1398](https://github.com/ehrbase/ehrbase/pull/1398))
+
+## [2.7.0]
+ ### Added
 * Experimental ItemTag REST endpoints for EHR_STATUS and COMPOSITION (configs: `ehrbase.rest.experimental.tags.*`) ([1343](https://github.com/ehrbase/ehrbase/pull/1343))
+* CLI runner with support for flyway pre-migrations ([1387](https://github.com/ehrbase/ehrbase/pull/1387))
  ### Changed 
  ### Fixed 
 * Require EHR_STATUS `is_queryable` and `is_modifiable` to be present ([#1377](https://github.com/ehrbase/ehrbase/pull/1377)) 
@@ -83,4 +112,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [2.4.0]: https://github.com/ehrbase/ehrbase/compare/v2.3.0...v2.4.0
 [2.5.0]: https://github.com/ehrbase/ehrbase/compare/v2.4.0...v2.5.0
 [2.6.0]: https://github.com/ehrbase/ehrbase/compare/v2.5.0...v2.6.0
-[unreleased]: https://github.com/ehrbase/ehrbase/compare/v2.6.0...HEAD
+[2.7.0]: https://github.com/ehrbase/ehrbase/compare/v2.6.0...v2.7.0
+[2.8.0]: https://github.com/ehrbase/ehrbase/compare/v2.7.0...v2.8.0
+[2.8.1]: https://github.com/ehrbase/ehrbase/compare/v2.8.0...v2.8.1
+[2.9.0]: https://github.com/ehrbase/ehrbase/compare/v2.8.1...v2.9.0
+[unreleased]: https://github.com/ehrbase/ehrbase/compare/v2.9.0...HEAD